requirements.txt
Flask>=2.3.0
opencv-python>=4.5.0
dlib>=19.22.0
numpy>=1.19.0
Pillow>=8.0.0
Werkzeug>=2.3.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
